+++ Source/BackStreet/Item/public/ProjectileInfoStruct.h
@@ -1,0 +1,47 @@
+#pragma once
+
+#include "Engine/DataTable.h"
+#include "../../Character/public/CharacteInfoEnum.h"
+#include "ProjectileInfoStruct.generated.h"
+
+
+USTRUCT(BlueprintType)
+struct FProjectileStatStruct : public FTableRowBase
+{
+public:
+	GENERATED_USTRUCT_BODY()
+
+	//발사체의 ID
+	UPROPERTY(VisibleDefaultsOnly, BlueprintReadOnly)
+		uint8 ProjectileID;
+
+	UPROPERTY(VisibleDefaultsOnly, BlueprintReadWrite)
+		FName ProjectileName;
+
+	UPROPERTY(VisibleDefaultsOnly, BlueprintReadWrite)
+		FName Description;
+
+	//폭발 타입의 발사체인지? (RadialDamage)
+	UPROPERTY(EditDefaultsOnly, BlueprintReadOnly)
+		bool bIsExplosive;
+
+	//발사체의 속도
+	UPROPERTY(EditDefaultsOnly, BlueprintReadOnly)
+		float ProjectileSpeed = 2000.0f;
+
+	//발사체의 데미지
+	UPROPERTY(EditDefaultsOnly, BlueprintReadOnly)
+		float ProjectileDamage = 0.2f;
+
+	//중력 스케일
+	UPROPERTY(EditDefaultsOnly, BlueprintReadOnly)
+		float GravityScale = 1.0f;
+
+	//유도가 되는지?
+	UPROPERTY(EditDefaultsOnly, BlueprintReadOnly)
+		bool bIsHoming = false;
+
+	//발사체는 각 하나의 디버프만 가짐
+	UPROPERTY(EditDefaultsOnly, BlueprintReadOnly)
+		ECharacterDebuffType DebuffType;
+};
